004 api search

+++ API-Search/Application/DTOs/Abstractions/PagedResultDto.cs
@@ -3,8 +3,8 @@
 public class PagedResultDto<T>
 {
     public IEnumerable<T> Items { get; set; } = Enumerable.Empty<T>();
-    public int TotalItems { get; set; }
+    public long TotalItems { get; set; }
     public int PageNumber { get; set; }
     public int PageSize { get; set; }
-    public long TotalPages { get; set; }
+    public int TotalPages { get; set; }
 }
